warrior callback

+++ Assets/_Scripts/CallbackObjects/AccountObject.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 
 public class AccountObject : CallbackObject {
 
-    public long id;
+    public int id;
     public string name;
 
     public AccountObject(){}
 
-    public AccountObject(long i, string s)
+    public AccountObject(int i, string s)
     {
         id = i;
         name = s;
